git2news: remember which commits have been posted.
git2news now tracks the IDs of commits that have already been posted
in a .seen file.  This file is a binary tree containing the IDs of all
commits so far posted.  It relies on the uniform distribution of
commit IDs to remain reasonably balanced.

The .seen file uses fixed-size bigendian values for its internal
offsets so it should be safe to transport between (reasonable)
architectures.